pyproject.toml
[project]
name = "duo-talk-evaluation"
version = "0.1.0"
description = "AI character dialogue quality evaluation system"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=0.2.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "pandas>=2.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
